fix(ci): make the merged main releasable — green gates + 0.3.0 release notes
The campaign merges left main's CI red with reconciliation debris; this
change fixes every failing job and expands the release notes so v0.3.0 can
be tagged:

- fmt: rustfmt over the hand-merged files
- clippy: DynamoRepository::count_active_sessions no longer holds the stats
  cache guard across the scan (re-stamp election, per the committed
  cache-lock discipline); test-only allow for the capture-gate test that
  deliberately holds the gate; drop a redundant cast
- test: the ffi embedder test allows its deliberate use of the deprecated
  synchronous entry point (deny-warnings builds treated it as an error)
- signing-path policy: protected root bumped to oidc-exchange-adapters@0.3.0
  (the workspace version bump had orphaned the 0.2.0 pin)
- nodejs/web-apps jobs: corepack enable after setup-node so the bare pnpm
  script steps resolve through the packageManager pin
- python/conformance jobs: uv sync --group build so the pinned maturin the
  develop/build steps invoke is actually installed
- python GIL-release test: the typed config now requires an https webhook
  URL, so the slow plain-HTTP webhook fixture became a stalling TLS
  listener bounded by the webhook client timeout; the internal shared
  secret meets the 32-byte floor
- nodejs test file formatted (oxfmt)
- RELEASE_NOTES.md: 0.3.0 section expanded to cover everything since
  v0.2.0 — breaking defaults (exchange-only role, two-listener admin plane,
  named auth mechanisms, typed fail-closed config, strict grant binding,
  rotating refresh tokens), the security-hardening set, and binding API
  changes

+++ crates/adapters/src/shared/token_endpoint.rs
@@ -462,9 +462,10 @@ mod tests {
 
         Mock::given(method("POST"))
             .and(path("/oauth/token"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(
-                "p".repeat(crate::shared::http::MAX_UPSTREAM_BODY_BYTES as usize + 1),
-            ))
+            .respond_with(
+                ResponseTemplate::new(200)
+                    .set_body_string("p".repeat(crate::shared::http::MAX_UPSTREAM_BODY_BYTES + 1)),
+            )
             .expect(1)
             .mount(&server)
             .await;
